+++ packages/core/src/refs/object.ts
@@ -2,19 +2,24 @@
 import { OutputRef, outputShapeKey } from '../types';
 import BaseTypeRef from './base';
 
-import { InterfaceParam, ObjectTypeOptions, SchemaTypes } from '..';
+import { InterfaceParam, ObjectTypeOptions, parentShapeKey, SchemaTypes } from '..';
 
-export default class ObjectRef<T> extends BaseTypeRef implements OutputRef {
+export default class ObjectRef<T, P = T> extends BaseTypeRef implements OutputRef {
   kind = 'Object' as const;
 
   [outputShapeKey]: T;
+  [parentShapeKey]: P;
 
   constructor(name: string) {
     super('Object', name);
   }
 }
 
-export class ImplementableObjectRef<Types extends SchemaTypes, Shape> extends ObjectRef<Shape> {
+export class ImplementableObjectRef<
+  Types extends SchemaTypes,
+  Shape,
+  Parent = Shape
+> extends ObjectRef<Shape, Parent> {
   private builder: GiraphQLSchemaTypes.SchemaBuilder<Types>;
 
   constructor(builder: GiraphQLSchemaTypes.SchemaBuilder<Types>, name: string) {
@@ -24,7 +29,7 @@ export class ImplementableObjectRef<Types extends SchemaTypes, Shape> extends Ob
   }
 
   implement<Interfaces extends InterfaceParam<Types>[]>(
-    options: ObjectTypeOptions<Types, ObjectRef<Types>, Shape, Interfaces>,
+    options: ObjectTypeOptions<Types, ObjectRef<Types>, Parent, Interfaces>,
   ) {
     return this.builder.objectType(this, options);
   }