+++ packages/core/tests/examples/patterns/builder.ts
@@ -1,5 +1,15 @@
 import SchemaBuilder from '../../../src';
 
-const builder = new SchemaBuilder({});
+export interface SchemaTypes {
+  Scalars: {
+    ID: {
+      Input: string;
+      Output: string;
+    };
+  };
+}
+export type TypesWithDefault = GiraphQLSchemaTypes.ExtendDefaultTypes<SchemaTypes>;
+
+const builder = new SchemaBuilder<SchemaTypes>({});
 
 export default builder;