+++ packages/core/tests/examples/patterns/schema.ts
@@ -1,4 +1,5 @@
 import './types/circular';
+import './types/multiple';
 import builder from './builder';
 
 const schema = builder.toSchema({});