+++ packages/deno/packages/plugin-dataloader/schema-builder.ts
@@ -1,11 +1,11 @@
 // @ts-nocheck
 import DataLoader from 'https://cdn.skypack.dev/dataloader?dts';
 import { GraphQLResolveInfo } from 'https://cdn.skypack.dev/graphql?dts';
-import SchemaBuilder, { createContextCache, FieldRef, InterfaceRef, SchemaTypes, } from '../core/index.ts';
+import SchemaBuilder, { createContextCache, FieldRef, InterfaceParam, InterfaceRef, SchemaTypes, } from '../core/index.ts';
 import { DataloaderObjectTypeOptions, LoadableNodeOptions } from './types.ts';
 import { LoadableObjectRef } from './util.ts';
 const schemaBuilderProto = SchemaBuilder.prototype as GiraphQLSchemaTypes.SchemaBuilder<SchemaTypes>;
-schemaBuilderProto.loadableObject = function loadableObject<Shape extends object, Key extends bigint | number | string, CacheKey = Key>(name: string, { load, loaderOptions, ...options }: DataloaderObjectTypeOptions<SchemaTypes, Shape, Key, CacheKey>) {
+schemaBuilderProto.loadableObject = function loadableObject<Shape extends object, Key extends bigint | number | string, Interfaces extends InterfaceParam<SchemaTypes>[], CacheKey = Key>(name: string, { load, loaderOptions, ...options }: DataloaderObjectTypeOptions<SchemaTypes, Shape, Key, Interfaces, CacheKey>) {
     const getDataloader = createContextCache((context: SchemaTypes["Context"]) => new DataLoader<Key, Shape, CacheKey>((keys: readonly Key[]) => (load as (keys: readonly Key[], context: SchemaTypes["Context"]) => Promise<Shape[]>)(keys, context), loaderOptions));
     const ref = new LoadableObjectRef<SchemaTypes, Shape, Shape, Key, CacheKey>(this, name, getDataloader);
     ref.implement({
@@ -13,11 +13,11 @@ schemaBuilderProto.loadableObject = function loadableObject<Shape extends object
         extensions: {
             getDataloader,
         },
-    });
+    } as never);
     return ref;
 };
 const TloadableNode = schemaBuilderProto.loadableNode;
-schemaBuilderProto.loadableNode = (function loadableNode<Shape extends object, Key extends bigint | number | string, CacheKey = Key>(this: GiraphQLSchemaTypes.SchemaBuilder<SchemaTypes>, name: string, { load, loaderOptions, ...options }: LoadableNodeOptions<SchemaTypes, Shape, Key, CacheKey>) {
+schemaBuilderProto.loadableNode = (function loadableNode<Shape extends object, Key extends bigint | number | string, Interfaces extends InterfaceParam<SchemaTypes>[], CacheKey = Key>(this: GiraphQLSchemaTypes.SchemaBuilder<SchemaTypes>, name: string, { load, loaderOptions, ...options }: LoadableNodeOptions<SchemaTypes, Shape, Key, Interfaces, CacheKey>) {
     if (typeof (this as GiraphQLSchemaTypes.SchemaBuilder<SchemaTypes> & Record<string, unknown>)
         .nodeInterfaceRef !== "function") {
         throw new TypeError("builder.loadableNode requires @giraphql/plugin-relay to be installed");
@@ -30,13 +30,13 @@ schemaBuilderProto.loadableNode = (function loadableNode<Shape extends object, K
             (this as GiraphQLSchemaTypes.SchemaBuilder<SchemaTypes> & {
                 nodeInterfaceRef: () => InterfaceRef<unknown>;
             }).nodeInterfaceRef(),
-        ] as never[],
+        ],
         loadMany: (ids: Key[], context: SchemaTypes["Context"]) => getDataloader(context).loadMany(ids),
         extensions: {
             getDataloader,
         },
     };
-    ref.implement(extendedOptions);
+    ref.implement(extendedOptions as never);
     this.configStore.onTypeConfig(ref, (nodeConfig) => {
         this.objectField(ref, "id", (t) => ((t as unknown) as {
             globalID: (options: Record<string, unknown>) => FieldRef<unknown>;