+++ packages/plugin-simple-objects/tests/examples/simple/builder.ts
@@ -2,16 +2,16 @@ import SchemaBuilder from '@giraphql/core';
 import SimpleObjectsPlugin from '../../../src';
 import { ContextType } from './types';
 
-type Types = {
+interface Types {
   Context: ContextType;
   Scalars: {
     ID: {
       Input: string;
-      Output: string | number;
+      Output: number | string;
     };
   };
-};
+}
 
 export default new SchemaBuilder<Types>({
-  plugins: [SimpleObjectsPlugin]
+  plugins: [SimpleObjectsPlugin],
 });